finally got dead working - pretty sure

+++ AntBot-GA/Team.h
@@ -26,7 +26,7 @@
 @property (nonatomic) float powerReturnShift;
 @property (nonatomic) float powerReturnSigma;
 @property (nonatomic) float chargeActiveSigma;
-@property (nonatomic) int casualties;
+@property (nonatomic) float casualties;
 //////////////POWER STUFF///////////////
 
 //Non-evolved variables:
